seqの大きさ変更
git-svn-id: https://my.honestyworks.jp/svn/private/t2cms_impl/trunk@702 50c3a9c6-9719-4a2f-96f6-0abb51d135c3

+++ src/migration/teeple2cms/003-meta_record.sql
@@ -1,1 +1,1 @@
-ALTER TABLE meta_record ADD seq TINYINT;
+ALTER TABLE meta_record ADD seq INT(10);
